requirements.txt
setuptools==49.6.0
numpy==1.20.0
matplotlib==3.5.0
scanpy==1.9.1
jupyterlab==3.0.16
ipython==7.18.1
anndata==0.7.6
pandas==1.2.5
scikit-learn==0.24.2
scipy==1.6.2
seaborn==0.11.1
torchvision==0.8.2
torch==1.7.1
pyyaml==6.0
wandb==0.12.0
ml-collections==0.1.0
ott-jax==0.2.8
POT==0.8.2
leidenalg==0.8.10
black==22.8.0
protobuf~=3.19.0
